# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos",
        "description": "API REST para gestionar productos, categorías, etiquetas, pedidos y usuarios. Utiliza autenticación JWT: primero inicia sesión en POST /api/auth/login y luego envía el token en el header `Authorization: Bearer <token>`. El catálogo de productos es público; los pedidos y la gestión por rol exigen token.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://ense-ar-laravel.test",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Panel de administrador",
                "description": "Solo accesible con token JWT de un usuario con rol admin.",
                "operationId": "dc1b94ea7171857b36212ad34ca36d29",
                "responses": {
                    "200": {
                        "description": "Acceso concedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "¡Bienvenido al panel de Administrador! Todo funciona perfecto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o ausente"
                    },
                    "403": {
                        "description": "Rol sin permisos para acceder"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleado/pedidos": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Gestión de pedidos para empleados",
                "description": "Solo accesible con token JWT de un usuario con rol empleado.",
                "operationId": "5a38d1a535409972e5c6740e93c8c14d",
                "responses": {
                    "200": {
                        "description": "Acceso concedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Área de gestión de pedidos para Empleados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o ausente"
                    },
                    "403": {
                        "description": "Rol sin permisos para acceder"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cliente/perfil": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Perfil del cliente",
                "description": "Solo accesible con token JWT de un usuario con rol cliente.",
                "operationId": "3040500afc1969c395b909180089c123",
                "responses": {
                    "200": {
                        "description": "Acceso concedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Bienvenido a tu perfil de Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o ausente"
                    },
                    "403": {
                        "description": "Rol sin permisos para acceder"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT obtenido en POST /api/auth/login. Enviar como: Authorization: Bearer <token>"
            }
        }
    },
...
}